+++ packages/stream_core/lib/src/user/connect_user_details_request.dart
@@ -1,5 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 
+import 'user.dart';
+
 part 'connect_user_details_request.g.dart';
 
 @JsonSerializable(createFactory: false)
@@ -12,6 +14,18 @@ class ConnectUserDetailsRequest {
     this.name,
     this.custom,
   });
+
+  factory ConnectUserDetailsRequest.fromUser(
+    User user, {
+    bool includeDetails = true,
+  }) {
+    return ConnectUserDetailsRequest(
+      id: user.id,
+      name: includeDetails ? user.originalName : null,
+      image: includeDetails ? user.image : null,
+      custom: includeDetails ? user.custom : null,
+    );
+  }
 
   final String id;
   final String? image;